+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: d825319 - v9
+| Author: FJrodafo - Mon, 26 Aug 2024 19:56:34 +0100
+| 
 * Commit: 6570367 - v8
 | Author: FJrodafo - Tue, 30 Jul 2024 20:59:39 +0100
 |